requirements.txt
streamlit
pandas
numpy
sentence-transformers
faiss-cpu
transformers
torch
python-dotenv
accelerate
sentencepiece